+++ src/Option.php
@@ -8,6 +8,8 @@ declare(strict_types=1);
 
 namespace BlueMvc\Forms;
 
+use BlueMvc\Core\Collections\CustomItemCollection;
+use BlueMvc\Core\Traits\CustomItemsTrait;
 use BlueMvc\Forms\Interfaces\OptionInterface;
 use BlueMvc\Forms\Traits\BuildTagTrait;
 
@@ -19,6 +21,7 @@ use BlueMvc\Forms\Traits\BuildTagTrait;
 class Option implements OptionInterface
 {
     use BuildTagTrait;
+    use CustomItemsTrait;
 
     /**
      * Constructs the option.
@@ -35,6 +38,7 @@ class Option implements OptionInterface
         $this->isSelected = false;
         $this->customData = null;
         $this->isDisabled = false;
+        $this->customItems = new CustomItemCollection();
     }
 
     /**