+++ src/RadioButton.php
@@ -8,6 +8,8 @@ declare(strict_types=1);
 
 namespace BlueMvc\Forms;
 
+use BlueMvc\Core\Collections\CustomItemCollection;
+use BlueMvc\Core\Traits\CustomItemsTrait;
 use BlueMvc\Forms\Interfaces\RadioButtonInterface;
 use BlueMvc\Forms\Traits\BuildTagTrait;
 
@@ -19,6 +21,7 @@ use BlueMvc\Forms\Traits\BuildTagTrait;
 class RadioButton implements RadioButtonInterface
 {
     use BuildTagTrait;
+    use CustomItemsTrait;
 
     /**
      * Constructs the radio button.
@@ -36,6 +39,7 @@ class RadioButton implements RadioButtonInterface
         $this->isSelected = false;
         $this->customData = null;
         $this->isDisabled = false;
+        $this->customItems = new CustomItemCollection();
     }
 
     /**