+++ tests/RadioButtonTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace BlueMvc\Forms\Tests;
 
+use BlueMvc\Core\Collections\CustomItemCollection;
 use BlueMvc\Forms\RadioButton;
 use PHPUnit\Framework\TestCase;
 
@@ -173,5 +174,60 @@ class RadioButtonTest extends TestCase
         self::assertTrue($radioButton->isDisabled());
         self::assertSame('<input type="radio" value="foo" disabled>', $radioButton->getHtml());
         self::assertSame('<input type="radio" value="foo" disabled>', $radioButton->__toString());
+    }
+
+    /**
+     * Test getCustomItem method.
+     */
+    public function testGetCustomItem()
+    {
+        $radioButton = new RadioButton('foo', 'bar');
+
+        self::assertNull($radioButton->getCustomItem('Foo'));
+        self::assertNull($radioButton->getCustomItem('Bar'));
+        self::assertNull($radioButton->getCustomItem('Baz'));
+    }
+
+    /**
+     * Test setCustomItem method.
+     */
+    public function testSetCustomItem()
+    {
+        $radioButton = new RadioButton('foo', 'bar');
+        $radioButton->setCustomItem('Foo', 1234);
+        $radioButton->setCustomItem('Bar', true);
+
+        self::assertSame(1234, $radioButton->getCustomItem('Foo'));
+        self::assertTrue($radioButton->getCustomItem('Bar'));
+        self::assertNull($radioButton->getCustomItem('Baz'));
+    }
+
+    /**
+     * Test getCustomItems method.
+     */
+    public function testGetCustomItems()
+    {
+        $radioButton = new RadioButton('foo', 'bar');
+        $radioButton->setCustomItem('Bar', 0.0);
+        $radioButton->setCustomItem('Baz', 'Foo');
+
+        self::assertSame(['Bar' => 0.0, 'Baz' => 'Foo'], iterator_to_array($radioButton->getCustomItems()));
+    }
+
+    /**
+     * Test setCustomItems method.
+     */
+    public function testSetCustomItems()
+    {
+        $customItemCollection = new CustomItemCollection();
+        $customItemCollection->set('Foo', [1, 2]);
+        $customItemCollection->set('Baz', false);
+
+        $radioButton = new RadioButton('foo', 'bar');
+        $radioButton->setCustomItems($customItemCollection);
+
+        self::assertSame([1, 2], $radioButton->getCustomItem('Foo'));
+        self::assertNull($radioButton->getCustomItem('Bar'));
+        self::assertFalse($radioButton->getCustomItem('Baz'));
     }
 }